// libs/common/src/model/dto/slamnav/slamnav.description.ts
export enum slamnav_description {
  STATUS_IMU = 'IMU, Gyro 센서 데이터',
  STATUS_IMU_ACC = 'Gyro 속도',
  STATUS_IMU_GYR = 'IMU 가속도',
  STATUS_IMU_IMU = 'IMU 속도',
  STATUS_LIDAR = 'Lidar 연결 상태 및 데이터',
  STATUS_LIDAR_CONNECTION = 'Lidar 연결 상태',
  STATUS_LIDAR_PORT = 'Lidar 연결 포트',
  STATUS_LIDAR_SERIAL_NUMBER = 'Lidar 시리얼 넘버',
  STATUS_MOTOR = '모터 연결 상태 및 데이터',
  STATUS_MOTOR_CONNECTION = '모터 연결 상태',
  STATUS_MOTOR_CURRENT = '모터 전류',
  STATUS_MOTOR_STATUS = '모터 상태 8가지를 8bit 형태로 쪼개어 각 비트자리수가 0 혹은 1일때에 따라서 상태가 변함. 낮은비트 순서로 (READY, MODE ERROR, JAM ERROR, CURRENT ERROR, BIG ERROR, INPUT ERROR, POSITION ERROR, COLLISTION ERROR)를 나타내며, status 값이 0일때는 Motor Not ready, 1일때는 Motor Ready, 16일때는 Motor Big Error, 20일때는 Motor Big Error + Motor Jam Error 등으로 판단한다.',
  STATUS_MOTOR_TEMP = '모터 온도',
  STATUS_CONDITION = '로봇 위치추정 상태',
  STATUS_CONDITION_INLIER_ERROR = '위치추정 에러율',
  STATUS_CONDITION_INLIER_RATIO = '위치추정 정확도',
  STATUS_CONDITION_MAPPING_ERROR = '매핑 에러율',
  STATUS_CONDITION_MAPPING_RATIO = '매핑 정확도',
  STATUS_STATE = '로봇 상태 (충전, 도킹, 전원, 초기화)',
  STATUS_STATE_CHARGE = '충전 상태. 로봇이 SRV 타입인 경우, none, ready(충전 중) 상태만 존재함',
  STATUS_STATE_DOCK = '도킹 상태',
  STATUS_STATE_EMO = '비상전원스위치 상태. 스위치가 눌린경우(전원 차단) 값이 true',
  STATUS_STATE_LOCALIZATION = '위치초기화 상태. 위치초기화가 되지 않은 상태에서는 none, 위치초기화가 성공적으로 되었을 경우 good, 위치초기화에 실패하거나 도중에 위치를 잃어버렸을 경우 fail 값을 가진다',
  STATUS_STATE_POWER = '전원 상태. 로봇 모터단으로 전원이 인가되는 상태인 경우 true',
  STATUS_POWER = '로봇 전원 상태',
  STATUS_POWER_BATTERY_CURRENT = '배터리 전류',
  STATUS_POWER_BATTERY_IN = '배터리 입력전원',
  STATUS_POWER_BATTERY_OUT = '배터리 출력전원',
  STATUS_POWER_BATTERY_PERCENT = '배터리 출력전원(퍼센트)',
  STATUS_POWER_CHARGE_CURRENT = '충전 전류',
  STATUS_POWER_CONTACT_VOLTAGE = '충전 전압',
  STATUS_POWER_POWER = '전력',
  STATUS_POWER_TOTAL_POWER = '누적 전력',
  STATUS_SETTING = '로봇의 세팅 값',
  STATUS_SETTING_PLATFORM_TYPE = '로봇 타입',
  STATUS_MAP = '로봇 맵 상태',
  STATUS_MAP_NAME = '로봇 맵 이름',
  STATUS_MAP_STATUS = '로봇 맵 로딩 상태. 맵이 로딩되지 않은 경우 none, 맵이 로딩중일때(파일이 클 수록 길어짐) loading, 로딩이 완료되었을 떄 loaded 값을 지닌다',
  TIME = '메시지 발송 시간. ms 단위',
  STATUS_MOVE = '이동 현 상태',
  STATUS_MOVE_AUTO = '자율주행 이동 상태',
  STATUS_MOVE_DOCK = '도킹 이동 상태(미사용)',
  STATUS_MOVE_JOG = '조이스틱 이동 상태(미사용)',
  STATUS_MOVE_OBS = '주행 중 장애물 상태',
  STATUS_MOVE_PATH = '주행 경로요청 상태',
  STATUS_POSE = '로봇 글로벌좌표. 위치초기화가 good인 상태일때 유의미',
  STATUS_VELOCITY = '로봇 주행 속도',
  STATUS_GOAL_NODE = '주행 목표지점 정보 및 상태',
  STATUS_GOAL_ID = '주행 목표지점 노드 ID',
  STATUS_GOAL_NAME = '주행 목표지점 노드 이름',
  STATUS_GOAL_STATE = '주행 목표지점 이동 상태',
  STATUS_GOAL_XYZ = '주행 목표지점 글로벌 좌표',
  STATUS_CUR_NODE = '주행 현재지점 정보 및 상태',
  STATUS_CUR_ID = '주행 현재지점 노드 ID',
  STATUS_CUR_NAME = '주행 현재지점 노드 이름',
  STATUS_CUR_STATE = '주행 현재지점 이동 상태',
  STATUS_CUR_XYZ = '주행 현재지점 글로벌 좌표',
  CLOUD_LIDAR = '현재 라이다 데이터 배열',
  CLOUD_MAPPING = '맵 생성 중 누적되는 라이다 데이터 배열(누적된 전체 배열이 아님)',
  CLOUD_XYZ = '라이다 클라우드의 로봇기준 좌표',
  CLOUD_INTENSITY = '라이다 클라우드의 강도. 맵 생성 중 누적된 횟수에 비례하여 높은 값을 지님',
  PATH_XYZ = '주행 중 생성한 경로 포인트의 글로벌좌표',
}

export enum GoalStatus {
  none = 'none',
  move = 'move',
  complete = 'complete',
  fail = 'fail',
  obstacle = 'obstacle',
  cancel = 'cancel',
}

export enum MoveStatus {
  move = 'move',
  error = 'error',
  pause = 'pause',
  stop = 'stop',
  notReady = 'not ready',
  vir = 'vir',
}
export enum ObsStatus {
  none = 'none',
  far = 'far',
  near = 'near',
  vir = 'vir',
}
export enum PathStatus {
  none = 'none',
  reqPath = 'req_path',
  recevPath = 'recv_path',
}

export enum MapStatus {
  none = 'none',
  loading = 'loading',
  loaded = 'loaded',
}

export enum LocalizationState {
  none = 'none',
  good = 'good',
  fail = 'fail',
}

export enum ChargeState {
  none = 'none',
  ready = 'ready',
  battery_on = 'battery_on',
  charging = 'charging',
  finish = 'finish',
  fail = 'fail',
}